\name{getMaster}
\alias{getMaster}
\alias{loadMaster}

\title{
  Get \emph{master} HDMSe data
}

\description{
  TODO A concise (1-5 lines) description of what the function does.
}

\usage{
getMaster()
loadMaster()
}


\details{
  TODO If necessary, more details than the description above 
}


\references{
  Bond N. J., Shliaha P.V., Gatto L. and Lilley K.S., in prep.

  See the \code{synapter} vignette from the \code{synapter} package,
  available with \code{ysnapterGuide()} for a description of the
  underlying concepts and detailed description of the pipeline.
}

\author{
  Laurent Gatto <lg390@cam.ac.uk>
}



\seealso{
  \code{\link{ups25a}} and \code{\link{getHDMSeFinalPeptide}}
}

\examples{
loadMaster()
master
}

