\name{synapterdata-package}
\alias{synapterdata-package}
\alias{synapterdata}
\docType{package}

\title{
Data accompanying the synapter package
}

\description{
Data independant acquisition of UPS1 protein mix in 
an \emph{E. coli} background obtained on a Waters Synapt G2 
instrument.
}

\details{
  See the \code{synapter} package for details.
  
  Index:
  \preformatted{
    getHDMSeFinalPeptide    PLGS csv data and fasta files
    getMaster               Get _master_ HDMSe data
    ups25a                  'Synapter' spiked-in example data.
  }
}

\author{
Laurent Gatto and Pavel V. Shliaha

Maintainer: Laurent Gatto <lg390@cam.ac.uk>
}

\references{
  Shliaha P.V.,  Gatto L., Bond N.J. and Lilley K.S.
  Synapter: Improving qualitative and quantitative performance for label free
  proteomics, in prep. 

  Shliaha, P.V., Bond N.J., Gatto L. and Lilley K.S.
  The Effects of Ion Mobility Separation on Data Independent Acquisition
  in Proteomics Studies., in prep. 
}

\keyword{ package }
