\name{synobj2}
\alias{synobj2}
\alias{synobj3}
\docType{data}

\title{
  Example Synapter object
}

\description{

  An instance of class \code{Synapter} used to demonstrate ion mobility
  separation and peak matching in the respective synapter vignettes. The
  object can be created using the files available at
  \url{http://proteome.sysbiol.cam.ac.uk/lgatto/synapter/data/} and the
  script \code{create_synobj2.R} in \code{system.file("scripts", package
  = "synapterdata")}.

  The \code{synobj3} object replaces \code{synobj2} for synapter
  version >= 1.99.

}

\usage{
data(synobj2)
data(synobj3)
}

\format{
  An instance of class \code{Synapter}.
}


\examples{
data(synobj2)
}

\keyword{datasets}
